Adding condensation of IH and TC into foundation.dart, and added casing specifications to IH
