+++ Codebase/Aureus-Package/lib/Elements/Cards/BadgeCardElement.dart
@@ -1,5 +1,3 @@
-import 'package:aureus/Elements/Info%20Hiearchy/InfoHierarchyElements.dart';
-import 'package:aureus/TitleCase.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:aureus/foundation.dart';
 import 'package:flutter/material.dart';
