+++ Codebase/Aureus-Package/lib/Elements/Cards/StandardCardElement.dart
@@ -1,5 +1,3 @@
-import 'package:aureus/Elements/Info%20Hiearchy/InfoHiearchyElements.dart';
-import 'package:aureus/Elements/Misc/TitleCase.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:aureus/foundation.dart';
 import 'package:flutter/material.dart';
