+++ Codebase/Aureus-Package/lib/foundation.dart
@@ -178,3 +178,85 @@ class UDSVariables {
     );
   }
 }
+
+class TitleCase {
+  static String convertToTitleCase(String text) {
+    if (text.length <= 1) {
+      return text.toUpperCase();
+    }
+
+    final List<String> words = text.split(' ');
+    final capitalizedWords = words.map((word) {
+      if (word.trim().isNotEmpty) {
+        final String firstLetter = word.trim().substring(0, 1).toUpperCase();
+        final String remainingLetters = word.trim().substring(1);
+
+        return '$firstLetter$remainingLetters';
+      }
+      return '';
+    });
+
+    return capitalizedWords.join(' ');
+  }
+}
+
+//Info Hiearchy - A set of texts / label styles that acts as the typographic foundation for the rest of the UDS
+
+final foundation = UDSVariables();
+final titleCase = TitleCase();
+
+class HeadingOneText extends Text {
+  HeadingOneText(String data, Color color)
+      : super(TitleCase.convertToTitleCase(data), style: foundation.heading1().copyWith(color: color));
+}
+
+class HeadingTwoText extends Text {
+  HeadingTwoText(String data, Color color)
+      : super(data.toUpperCase(), style: foundation.heading2().copyWith(color: color));
+}
+
+class HeadingThreeText extends Text {
+  HeadingThreeText(String data, Color color)
+      : super(data.toUpperCase(), style: foundation.heading3().copyWith(color: color));
+}
+
+class HeadingFourText extends Text {
+  HeadingFourText(String data, Color color)
+      : super(data.toUpperCase(), style: foundation.heading4().copyWith(color: color));
+}
+
+class SubheaderText extends Text {
+  SubheaderText(String data, Color color)
+      : super(TitleCase.convertToTitleCase(data), style: foundation.subheading().copyWith(color: color));
+}
+
+class BodyOneText extends Text {
+  BodyOneText(String data, Color color)
+      : super(data, style: foundation.body1().copyWith(color: color));
+}
+
+class BodyTwoText extends Text {
+  BodyTwoText(String data, Color color)
+      : super(data, style: foundation.body2().copyWith(color: color));
+}
+
+class ButtonOneText extends Text {
+  ButtonOneText(String data, Color color)
+      : super(data.toUpperCase(), style: foundation.button1().copyWith(color: color));
+}
+
+class ButtonTwoText extends Text {
+  ButtonTwoText(String data, Color color)
+      : super(data.toUpperCase(), style: foundation.button2().copyWith(color: color));
+}
+
+class TagOneText extends Text {
+  TagOneText(String data, Color color)
+      : super(data.toUpperCase(), style: foundation.tag1().copyWith(color: color));
+}
+
+class TagTwoText extends Text {
+  TagTwoText(String data, Color color)
+      : super(data.toUpperCase(), style: foundation.tag2().copyWith(color: color));
+}
+
